Added Hackney contract models
